stremio_core_web::model::serialize_meta_details
Function serialize_meta_details

pub fn serialize_meta_details<E: Env + 'static>(
    meta_details: &MetaDetails,
    ctx: &Ctx,
    streaming_server: &StreamingServer,
) -> JsValue

For MetaDetails:

If at least 1 item is ready we show the first ready item’s data

If all loaded resources have returned an error we show the first item’s error

We show a loading state